Simplify by using EntryType
