+++ src/app/search/search-entry-table.component.html
@@ -1,0 +1,121 @@
+<div>
+  <div *ngIf="!dataSource">
+    <mat-progress-bar mode="indeterminate"></mat-progress-bar>
+  </div>
+  <div fxLayout="row" fxLayoutAlign="space-between center">
+    <mat-form-field appearance="outline" class="search-form-field sort-by">
+      <mat-label>Sort by</mat-label>
+      <mat-select [value]="defaultSortOption.sort" (selectionChange)="setSort($event.value)">
+        <mat-option *ngFor="let sortOption of sortOptions" [value]="sortOption.sort">{{ sortOption.label }}</mat-option>
+      </mat-select>
+    </mat-form-field>
+    <div class="tag-cloud-dropdown">
+      <button
+        mat-button
+        appearance="outline"
+        fxFlex="21rem"
+        class="tagCloud-btn {{ entryType | lowercase }}-background"
+        (click)="clickTagCloudBtn(this.entryType)"
+      >
+        <div fxLayout="row" fxLayoutAlign="space-between center">
+          <div>
+            <img class="tag-cloud" src="../assets/images/dockstore/tag-cloud.svg" alt="Popular keywords icon" />
+            <span class="m-2"><b>Popular Keywords</b></span>
+          </div>
+          <fa-icon [icon]="showTagCloudForEntryType ? faMinus : faPlus" aria-hidden="true"></fa-icon>
+        </div>
+      </button>
+      <div class="tagCloud {{ entryType | lowercase }}" *ngIf="showTagCloudForEntryType && tagCloudData" data-cy="tagCloud">
+        <angular-tag-cloud
+          [data]="tagCloudData"
+          [width]="options.width"
+          [height]="options.height"
+          [overflow]="options.overflow"
+          (clicked)="tagClicked($event)"
+        >
+        </angular-tag-cloud>
+      </div>
+    </div>
+  </div>
+  <mat-table [hidden]="!dataSource" [dataSource]="dataSource" matSort data-cy="search-entry-results-table" multiTemplateDataRows>
+    <ng-container matColumnDef="name">
+      <mat-header-cell *matHeaderCellDef></mat-header-cell>
+      <mat-cell data-cy="entryColumn" *matCellDef="let entry">
+        <mat-card fxLayout="column" fxLayoutGap="0.3rem" class="w-100">
+          <div fxLayout="row" fxLayoutAlign="space-between">
+            <div>
+              <img
+                src="../../../../assets/svg/{{ entry?.source.entryTypeMetadata.term }}-circle.svg"
+                class="mr-2_5 site-icons-small"
+                alt="{{ entry?.source.entryTypeMetadata.term }} icon"
+              />
+              <a
+                [matTooltip]="entry?.source | entryToDisplayName: true"
+                matTooltipPosition="left"
+                [routerLink]="entry?.source | routerLink"
+                >{{ entry?.source | entryToDisplayName: true }}</a
+              >
+            </div>
+            <div class="star-color weight-bold">
+              <mat-icon class="mat-icon-sm" *ngIf="entry?.source.starredUsers?.length > 0">star_rate</mat-icon>
+              <span>{{
+                !entry?.source.starredUsers || entry?.source.starredUsers.length === 0 ? '' : entry?.source.starredUsers.length
+              }}</span>
+            </div>
+          </div>
+          <div>
+            <span>{{ entry?.source.topicAutomatic }}</span>
+            <app-ai-bubble
+              *ngIf="entry?.source.topicSelection === TopicSelectionEnum.AI && !entry?.source.approvedAITopic"
+              [isPublic]="true"
+            ></app-ai-bubble>
+          </div>
+          <!-- Fix this in the webservice so that it's not possible to have an array with a null value -->
+          <div *ngIf="entry?.source?.all_authors" class="truncate-text-3 size-small">
+            <mat-icon class="info-icon">account_circle</mat-icon
+            ><span [innerHTML]="entry?.source?.all_authors | getSearchAuthorsHtml"></span>
+          </div>
+
+          <!-- Search highlighting results-->
+          <div *ngIf="entry.highlight" class="my-4">
+            <table class="size-small">
+              <tr *ngFor="let highlight of entry.highlight | keyvalue">
+                <th>{{ this.searchEverythingFriendlyNames.get(highlight.key.toString()) }}:</th>
+                <td><span [innerHTML]="highlight.value | joinWithEllipses"></span></td>
+              </tr>
+            </table>
+          </div>
+
+          <div fxLayout="row" fxLayoutAlign="space-between">
+            <div>
+              <span class="size-small mr-2"
+                >Last updated {{ entry?.source.last_modified_date ? (entry?.source.last_modified_date | date) : 'n/a' }}</span
+              >
+              <span *ngIf="entry?.source.entryTypeMetadata.type !== EntryType.TOOL" class="nonclick-badge">{{
+                entry?.source.descriptorType | descriptorLanguage
+              }}</span>
+              <span *ngIf="entry?.source.entryTypeMetadata.type === EntryType.TOOL">
+                <span *ngFor="let language of entry?.source.descriptorType" class="nonclick-badge m-1">
+                  {{ language ? language : '' }}
+                </span>
+              </span>
+            </div>
+            <app-doi-badge
+              *ngIf="entry?.source?.selected_concept_doi"
+              [doi]="entry.source.selected_concept_doi"
+              [displayInitiator]="false"
+            ></app-doi-badge>
+          </div>
+        </mat-card>
+      </mat-cell>
+    </ng-container>
+
+    <mat-row *matRowDef="let element; columns: displayedColumns" class="element-row py-2"></mat-row>
+  </mat-table>
+  <mat-paginator
+    data-cy="search-entry-table-paginator"
+    [showFirstLastButtons]="true"
+    [pageSizeOptions]="[10, 20, 50, 100]"
+    (page)="updatePageSizeAndIndex($event)"
+  ></mat-paginator>
+</div>
