+++ src/app/search/search-entry-table.component.scss
@@ -1,5 +1,5 @@
 /*!
- *    Copyright 2017 OICR
+ *    Copyright 2025 OICR
  *
  *    Licensed under the Apache License, Version 2.0 (the "License");
  *    you may not use this file except in compliance with the License.
@@ -15,3 +15,30 @@
  */
 @use '@angular/material' as mat;
 @import '/src/materialColorScheme.scss';
+
+.tag-cloud-dropdown {
+  position: relative;
+  display: inline-block !important;
+}
+
+.tagCloud {
+  position: absolute;
+  right: 0;
+  top: 3.5rem;
+  z-index: 1;
+  border-radius: 4px;
+  background-color: #fff;
+  box-shadow: 0 0 10px 0px mat.get-color-from-palette($dockstore-app-gray, 7);
+}
+
+.tagCloud-btn {
+  box-shadow: 1px 3px 5px 0px mat.get-color-from-palette($dockstore-app-gray, 7);
+  color: $header-color;
+}
+
+::ng-deep .sort-by {
+  .mat-form-field-wrapper {
+    // Removing the padding because we don't need hint
+    padding-bottom: 0;
+  }
+}
