--- src/app/search/search-notebook-table/search-notebook-table.component.html
@@ -1,149 +1,0 @@
-<div>
-  <div *ngIf="!dataSource">
-    <mat-progress-bar mode="indeterminate"></mat-progress-bar>
-  </div>
-  <mat-form-field appearance="outline" class="search-form-field">
-    <mat-label>Sort by</mat-label>
-    <mat-select [value]="defaultSortOption.sort" (selectionChange)="setSort($event.value)">
-      <mat-option *ngFor="let sortOption of sortOptions" [value]="sortOption.sort">{{ sortOption.label }}</mat-option>
-    </mat-select>
-  </mat-form-field>
-  <mat-table [hidden]="!dataSource" [dataSource]="dataSource" matSort data-cy="search-notebook-results-table" multiTemplateDataRows>
-    <ng-container matColumnDef="name">
-      <mat-header-cell *matHeaderCellDef></mat-header-cell>
-      <mat-cell data-cy="notebookColumn" *matCellDef="let notebook">
-        <mat-card fxLayout="column" fxLayoutGap="0.3rem" class="w-100">
-          <div fxLayout="row" fxLayoutAlign="space-between">
-            <div>
-              <img src="../../../../assets/svg/notebook-circle.svg" class="mr-2_5 site-icons-small" alt="notebook icon" />
-              <a
-                [matTooltip]="notebook?.source.full_workflow_path"
-                matTooltipPosition="left"
-                [routerLink]="'/notebooks/' + notebook?.source.full_workflow_path"
-                >{{
-                  notebook?.source.organization +
-                    '/' +
-                    notebook?.source.repository +
-                    (notebook?.source.workflowName ? '/' + notebook?.source.workflowName : '')
-                }}</a
-              >
-            </div>
-            <div class="star-color weight-bold">
-              <mat-icon class="mat-icon-sm" *ngIf="notebook?.source.starredUsers?.length > 0">star_rate</mat-icon>
-              <span>{{
-                !notebook?.source.starredUsers || notebook?.source.starredUsers.length === 0 ? '' : notebook?.source.starredUsers.length
-              }}</span>
-            </div>
-          </div>
-          <div>
-            <span>{{ notebook?.source.topicAutomatic }}</span>
-            <app-ai-bubble
-              *ngIf="notebook?.source.topicSelection === TopicSelectionEnum.AI && !notebook?.source.approvedAITopic"
-              [isPublic]="true"
-            ></app-ai-bubble>
-          </div>
-          <!-- Fix this in the webservice so that it's not possible to have an array with a null value -->
-          <div *ngIf="notebook?.source?.all_authors" class="truncate-text-3 size-small">
-            <mat-icon class="info-icon">account_circle</mat-icon
-            ><span [innerHTML]="notebook?.source?.all_authors | getSearchAuthorsHtml"></span>
-          </div>
-
-          <!-- Search highlighting results-->
-          <div *ngIf="notebook.highlight" class="my-4">
-            <table class="size-small">
-              <tr *ngFor="let highlight of notebook.highlight | keyvalue">
-                <th>{{ this.searchEverythingFriendlyNames.get(highlight.key.toString()) }}:</th>
-                <td><span [innerHTML]="highlight.value | joinWithEllipses"></span></td>
-              </tr>
-            </table>
-          </div>
-
-          <div fxLayout="row" fxLayoutAlign="space-between">
-            <div>
-              <span class="size-small mr-2"> Last updated {{ notebook?.source.last_modified_date | date }} </span>
-              <span class="nonclick-badge">{{ notebook?.source.descriptorType | descriptorLanguage }}</span>
-            </div>
-            <app-doi-badge
-              *ngIf="notebook?.source?.selected_concept_doi"
-              [doi]="notebook.source.selected_concept_doi"
-              [displayInitiator]="false"
-            ></app-doi-badge>
-          </div>
-        </mat-card>
-
-        <!-- <div>
-          <img src="../../../../assets/svg/notebook-circle.svg" class="mr-2_5 site-icons-small" alt="notebook icon" />
-          <a [matTooltip]="notebook?.source.full_workflow_path" [routerLink]="'/notebooks/' + notebook?.source.full_workflow_path">{{
-            notebook?.source.organization +
-              '/' +
-              notebook?.source.repository +
-              (notebook?.source.workflowName ? '/' + notebook?.source.workflowName : '')
-          }}</a>
-          <div>
-            <span>{{ notebook?.source.topicAutomatic }}</span>
-            <app-ai-bubble
-              *ngIf="notebook?.source.topicSelection === TopicSelectionEnum.AI && !notebook?.source.approvedAITopic"
-              [isPublic]="true"
-            ></app-ai-bubble>
-          </div>
-        </div> -->
-      </mat-cell>
-    </ng-container>
-    <ng-container matColumnDef="all_authors">
-      <mat-header-cell fxShow fxHide.lt-sm *matHeaderCellDef mat-sort-header>Author</mat-header-cell>
-      <mat-cell fxShow fxHide.lt-sm *matCellDef="let notebook" [matTooltip]="notebook?.source.all_authors | getSearchAuthorsHtml: false">
-        <div class="truncate-text-3 size-small">
-          <span [innerHTML]="notebook?.source?.all_authors | getSearchAuthorsHtml"></span>
-        </div>
-      </mat-cell>
-    </ng-container>
-    <ng-container matColumnDef="descriptorType">
-      <mat-header-cell fxShow fxHide.lt-sm data-cy="descriptorTypeHeader" *matHeaderCellDef mat-sort-header>Format</mat-header-cell>
-      <mat-cell fxShow fxHide.lt-sm data-cy="descriptorType" *matCellDef="let notebook">
-        <div>
-          <span class="nonclick-badge">{{ notebook?.source.descriptorType | descriptorLanguage }}</span>
-        </div>
-      </mat-cell>
-    </ng-container>
-    <ng-container matColumnDef="descriptorTypeSubclass">
-      <mat-header-cell fxShow fxHide.lt-sm data-cy="descriptorTypeSubclassHeader" *matHeaderCellDef mat-sort-header
-        >Language</mat-header-cell
-      >
-      <mat-cell fxShow fxHide.lt-sm data-cy="descriptorTypeSubclass" *matCellDef="let notebook">
-        <div>
-          <span class="nonclick-badge">{{ notebook?.source.descriptorTypeSubclass }}</span>
-        </div>
-      </mat-cell>
-    </ng-container>
-    <ng-container matColumnDef="starredUsers">
-      <mat-header-cell *matHeaderCellDef mat-sort-header start="desc">Stars</mat-header-cell>
-      <mat-cell class="starred-cell weight-bold" *matCellDef="let notebook"
-        ><mat-icon class="star-icon" *ngIf="notebook?.source.starredUsers?.length > 0">star_rate</mat-icon>
-        {{ !notebook?.source.starredUsers || notebook?.source.starredUsers.length === 0 ? '' : notebook?.source.starredUsers.length }}
-      </mat-cell>
-    </ng-container>
-    <!-- Search highlighting column - this row is made up of this one column that spans across all columns -->
-    <!-- <ng-container matColumnDef="searchHighlight">
-      <mat-cell *matCellDef="let notebook" [attr.colspan]="displayedColumns.length">
-        <div *ngIf="notebook.highlight" class="pb-3">
-          <table class="size-small">
-            <tr *ngFor="let highlight of notebook.highlight | keyvalue">
-              <th>{{ this.searchEverythingFriendlyNames.get(highlight.key.toString()) }}:</th>
-              <td><span [innerHTML]="highlight.value | joinWithEllipses"></span></td>
-            </tr>
-          </table>
-        </div>
-      </mat-cell>
-    </ng-container> -->
-
-    <!-- <mat-header-row *matHeaderRowDef="displayedColumns"></mat-header-row> -->
-    <mat-row *matRowDef="let row; columns: displayedColumns" class="element-row py-2"></mat-row>
-    <!-- <mat-row *matRowDef="let row; columns: ['searchHighlight']" class="element-row search-highlight-row"></mat-row> -->
-  </mat-table>
-  <mat-paginator
-    data-cy="search-notebook-table-paginator"
-    [showFirstLastButtons]="true"
-    [pageSizeOptions]="[10, 20, 50, 100]"
-    (page)="updatePageSizeAndIndex($event)"
-  ></mat-paginator>
-</div>
