--- src/app/search/search-notebook-table/search-notebook-table.component.scss
@@ -1,11 +1,0 @@
-.mat-column-descriptorType {
-  max-width: 10rem;
-}
-
-.mat-column-descriptorTypeSubclass {
-  max-width: 10rem;
-}
-
-td.mat-cell {
-  border-bottom-style: none;
-}
