--- src/app/search/search-tool-table/search-tool-table.component.html
@@ -1,115 +1,0 @@
-<!--
-  ~    Copyright 2018 OICR
-  ~
-  ~    Licensed under the Apache License, Version 2.0 (the "License");
-  ~    you may not use this file except in compliance with the License.
-  ~    You may obtain a copy of the License at
-  ~
-  ~        http://www.apache.org/licenses/LICENSE-2.0
-  ~
-  ~    Unless required by applicable law or agreed to in writing, software
-  ~    distributed under the License is distributed on an "AS IS" BASIS,
-  ~    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-  ~    See the License for the specific language governing permissions and
-  ~    limitations under the License.
-  -->
-<div>
-  <div *ngIf="!dataSource">
-    <mat-progress-bar mode="indeterminate"></mat-progress-bar>
-  </div>
-  <mat-table [hidden]="!dataSource" [dataSource]="dataSource" matSort multiTemplateDataRows>
-    <ng-container matColumnDef="name">
-      <mat-header-cell *matHeaderCellDef mat-sort-header>Name and Description</mat-header-cell>
-      <!-- the wrap causes slightly odd behavior with the mat-icon but is needed for the topic -->
-      <mat-cell data-cy="toolNames" *matCellDef="let tool">
-        <div>
-          <img src="../../../../assets/svg/tool-circle.svg" class="mr-2_5 site-icons-small" alt="tool icon" />
-          <span *ngIf="tool?.source.private_access">
-            <app-private-icon></app-private-icon>
-          </span>
-          <span *ngIf="tool.source | isAppTool; else docktoreToolName">
-            <a
-              [matTooltip]="tool?.source.full_workflow_path"
-              matTooltipPosition="left"
-              [routerLink]="'/containers/' + tool.source.full_workflow_path"
-              >{{
-                tool?.source.organization +
-                  '/' +
-                  tool?.source.repository +
-                  (tool?.source.workflowName ? '/' + tool?.source.workflowName : '')
-              }}</a
-            >
-          </span>
-          <ng-template #docktoreToolName>
-            <a [matTooltip]="tool?.source.tool_path" matTooltipPosition="left" [routerLink]="'/containers/' + tool.source.tool_path">{{
-              tool?.source.namespace + '/' + tool?.source.name + (tool?.source.toolname ? '/' + tool?.source.toolname : '')
-            }}</a>
-          </ng-template>
-          <div>
-            <span>{{ tool?.source.topicAutomatic }}</span>
-            <app-ai-bubble
-              *ngIf="tool?.source.topicSelection === TopicSelectionEnum.AI && !tool?.source.approvedAITopic"
-              [isPublic]="true"
-            ></app-ai-bubble>
-          </div>
-        </div>
-      </mat-cell>
-    </ng-container>
-    <ng-container matColumnDef="all_authors">
-      <mat-header-cell fxShow fxHide.lt-sm *matHeaderCellDef mat-sort-header>Author</mat-header-cell>
-      <mat-cell fxShow fxHide.lt-sm *matCellDef="let tool" [matTooltip]="tool?.source.all_authors | getSearchAuthorsHtml: false">
-        <div class="truncate-text-3 size-small">
-          <span [innerHTML]="tool?.source.all_authors | getSearchAuthorsHtml"></span>
-        </div>
-      </mat-cell>
-    </ng-container>
-    <ng-container matColumnDef="descriptorType" style="width: fit-content">
-      <mat-header-cell fxShow fxHide.lt-sm *matHeaderCellDef mat-sort-header fxFlex="15%">Format</mat-header-cell>
-      <mat-cell data-cy="toolTypes" fxShow fxHide.lt-sm *matCellDef="let tool" fxFlex="15%" class="align-languages">
-        <ng-container *ngIf="tool.source | isAppTool; else dockstoreToolDescriptorType">
-          <div class="m-1">
-            <span class="nonclick-badge">{{ tool?.source.descriptorType | uppercase }}</span>
-          </div>
-        </ng-container>
-        <ng-template #dockstoreToolDescriptorType>
-          <div *ngFor="let language of tool?.source.descriptorType" class="m-1">
-            <span class="nonclick-badge">
-              {{ language ? (language.toString() | uppercase) : '' }}
-            </span>
-          </div>
-        </ng-template>
-      </mat-cell>
-    </ng-container>
-    <ng-container matColumnDef="starredUsers">
-      <mat-header-cell *matHeaderCellDef mat-sort-header start="desc">Stars</mat-header-cell>
-      <mat-cell class="description-cell" *matCellDef="let tool"
-        ><mat-icon class="star-icon" *ngIf="tool?.source.starredUsers?.length > 0">star_rate</mat-icon>
-        {{ !tool?.source.starredUsers || tool?.source.starredUsers.length === 0 ? '' : tool?.source.starredUsers.length }}
-      </mat-cell>
-    </ng-container>
-
-    <!-- Search highlighting column - this row is made up of this one column that spans across all columns -->
-    <ng-container matColumnDef="searchHighlight">
-      <mat-cell *matCellDef="let tool" [attr.colspan]="displayedColumns.length">
-        <div *ngIf="tool.highlight" class="pb-3">
-          <table class="size-small">
-            <tr *ngFor="let highlight of tool.highlight | keyvalue">
-              <th>{{ this.searchEverythingFriendlyNames.get(highlight.key.toString()) }}:</th>
-              <td><span [innerHTML]="highlight.value | joinWithEllipses"></span></td>
-            </tr>
-          </table>
-        </div>
-      </mat-cell>
-    </ng-container>
-
-    <mat-header-row *matHeaderRowDef="displayedColumns"></mat-header-row>
-    <mat-row *matRowDef="let row; columns: displayedColumns" class="py-3 element-row"></mat-row>
-    <mat-row *matRowDef="let row; columns: ['searchHighlight']" class="search-highlight-row"></mat-row>
-  </mat-table>
-  <mat-paginator
-    data-cy="search-tool-table-paginator"
-    [showFirstLastButtons]="true"
-    [pageSizeOptions]="[10, 20, 50, 100]"
-    (page)="updatePageSizeAndIndex($event)"
-  ></mat-paginator>
-</div>
