--- src/app/search/search-workflow-table/search-workflow-table.component.html
@@ -1,93 +1,0 @@
-<div>
-  <div *ngIf="!dataSource">
-    <mat-progress-bar mode="indeterminate"></mat-progress-bar>
-  </div>
-  <mat-form-field appearance="outline" class="search-form-field">
-    <mat-label>Sort by</mat-label>
-    <mat-select [value]="defaultSortOption.sort" (selectionChange)="setSort($event.value)">
-      <mat-option *ngFor="let sortOption of sortOptions" [value]="sortOption.sort">{{ sortOption.label }}</mat-option>
-    </mat-select>
-  </mat-form-field>
-  <mat-divider></mat-divider>
-  <mat-table [hidden]="!dataSource" [dataSource]="dataSource" matSort data-cy="search-workflow-results-table" multiTemplateDataRows>
-    <ng-container matColumnDef="name">
-      <mat-header-cell *matHeaderCellDef></mat-header-cell>
-      <mat-cell data-cy="workflowColumn" *matCellDef="let workflow">
-        <mat-card fxLayout="column" fxLayoutGap="0.3rem" class="w-100">
-          <div fxLayout="row" fxLayoutAlign="space-between">
-            <div>
-              <img
-                src="../../../../assets/svg/{{ entryTypeMetadata.term }}-circle.svg"
-                class="mr-2_5 site-icons-small"
-                alt="{{ entryTypeMetadata.term }} icon"
-              />
-              <a
-                [matTooltip]="workflow?.source.full_workflow_path"
-                matTooltipPosition="left"
-                [routerLink]="workflow?.source | routerLink"
-                >{{ workflow?.source | entryToDisplayName: true }}</a
-              >
-            </div>
-            <div class="star-color weight-bold">
-              <mat-icon class="mat-icon-sm" *ngIf="workflow?.source.starredUsers?.length > 0">star_rate</mat-icon>
-              <span>{{
-                !workflow?.source.starredUsers || workflow?.source.starredUsers.length === 0 ? '' : workflow?.source.starredUsers.length
-              }}</span>
-            </div>
-          </div>
-          <div>
-            <span>{{ workflow?.source.topicAutomatic }}</span>
-            <app-ai-bubble
-              *ngIf="workflow?.source.topicSelection === TopicSelectionEnum.AI && !workflow?.source.approvedAITopic"
-              [isPublic]="true"
-            ></app-ai-bubble>
-          </div>
-          <!-- Fix this in the webservice so that it's not possible to have an array with a null value -->
-          <div *ngIf="workflow?.source?.all_authors" class="truncate-text-3 size-small">
-            <mat-icon class="info-icon">account_circle</mat-icon
-            ><span [innerHTML]="workflow?.source?.all_authors | getSearchAuthorsHtml"></span>
-          </div>
-
-          <!-- Search highlighting results-->
-          <div *ngIf="workflow.highlight" class="my-4">
-            <table class="size-small">
-              <tr *ngFor="let highlight of workflow.highlight | keyvalue">
-                <th>{{ this.searchEverythingFriendlyNames.get(highlight.key.toString()) }}:</th>
-                <td><span [innerHTML]="highlight.value | joinWithEllipses"></span></td>
-              </tr>
-            </table>
-          </div>
-
-          <div fxLayout="row" fxLayoutAlign="space-between">
-            <div>
-              <span class="size-small mr-2"
-                >Last updated {{ workflow?.source.last_modified_date ? (workflow?.source.last_modified_date | date) : 'n/a' }}</span
-              >
-              <span *ngIf="workflow?.source.entryTypeMetadata.type !== EntryType.TOOL" class="nonclick-badge">{{
-                workflow?.source.descriptorType | descriptorLanguage
-              }}</span>
-              <span *ngIf="workflow?.source.entryTypeMetadata.type === EntryType.TOOL">
-                <span *ngFor="let language of workflow?.source.descriptorType" class="nonclick-badge m-1">
-                  {{ language ? language : '' }}
-                </span>
-              </span>
-            </div>
-            <app-doi-badge
-              *ngIf="workflow?.source?.selected_concept_doi"
-              [doi]="workflow.source.selected_concept_doi"
-              [displayInitiator]="false"
-            ></app-doi-badge>
-          </div>
-        </mat-card>
-      </mat-cell>
-    </ng-container>
-
-    <mat-row *matRowDef="let element; columns: displayedColumns" class="element-row py-2"></mat-row>
-  </mat-table>
-  <mat-paginator
-    data-cy="search-workflow-table-paginator"
-    [showFirstLastButtons]="true"
-    [pageSizeOptions]="[10, 20, 50, 100]"
-    (page)="updatePageSizeAndIndex($event)"
-  ></mat-paginator>
-</div>
