--- src/app/search/search-workflow-table/search-workflow-table.component.scss
@@ -1,3 +1,0 @@
-.mat-column-descriptorType {
-  max-width: 10rem;
-}
