+++ src/app/search/search.component.html
@@ -40,21 +40,18 @@
       <img class="site-icons-tab m-2" src="../assets/svg/sub-nav/workflow.svg" alt="workflow icon" />
       <b>Workflows</b>
     </ng-template>
-    <div class="mat-tab-content"></div>
   </mat-tab>
   <mat-tab>
     <ng-template mat-tab-label>
       <img class="site-icons-tab m-2" src="../assets/svg/sub-nav/tool.svg" alt="tool icon" />
       <b>Tools</b>
     </ng-template>
-    <div class="mat-tab-content"></div>
   </mat-tab>
   <mat-tab>
     <ng-template mat-tab-label>
       <img class="site-icons-tab m-2" src="../../assets/svg/sub-nav/notebook.svg" alt="notebook icon" />
       <b>Notebooks</b>
     </ng-template>
-    <div class="mat-tab-content"></div>
   </mat-tab>
 </mat-tab-group>
