+++ src/app/search/state/search.service.spec.ts
@@ -20,7 +20,7 @@ import { RouterTestingModule } from '@angular/router/testing';
 import { first } from 'rxjs/operators';
 import { ImageProviderService } from '../../shared/image-provider.service';
 import { ProviderService } from '../../shared/provider.service';
-import { Workflow } from '../../shared/openapi';
+import { EntryType, Workflow } from '../../shared/openapi';
 import { elasticSearchResponse } from '../../test/mocked-objects';
 import { ProviderStubService } from '../../test/service-stubs';
 import { Hit, SearchService } from './search.service';
@@ -148,16 +148,16 @@ describe('SearchService', () => {
     const c: Workflow = { ...a, authors: [], full_workflow_path: null, descriptorType: Workflow.DescriptorTypeEnum.WDL };
     c['all_authors'] = c['authors'];
 
-    expect(searchService.compareAttributes(a, b, 'all_authors', 'asc', 'workflow')).toEqual(-1);
-    expect(searchService.compareAttributes(a, b, 'all_authors', 'desc', 'workflow')).toEqual(1);
-    expect(searchService.compareAttributes(b, c, 'all_authors', 'asc', 'workflow')).toEqual(-1);
-    expect(searchService.compareAttributes(b, c, 'all_authors', 'desc', 'workflow')).toEqual(1);
-    expect(searchService.compareAttributes(a, c, 'descriptorType', 'asc', 'workflow')).toEqual(-1);
-    expect(searchService.compareAttributes(a, b, 'descriptorType', 'desc', 'workflow')).toEqual(-0);
-    expect(searchService.compareAttributes(a, b, 'starredUsers', 'asc', 'workflow')).toEqual(-1);
-    expect(searchService.compareAttributes(a, b, 'name', 'asc', 'workflow')).toEqual(-1);
-    expect(searchService.compareAttributes(a, b, 'name', 'desc', 'workflow')).toEqual(1);
-    expect(searchService.compareAttributes(b, c, 'name', 'asc', 'workflow')).toEqual(-1);
-    expect(searchService.compareAttributes(b, c, 'name', 'desc', 'workflow')).toEqual(-1);
+    expect(searchService.compareAttributes(a, b, 'all_authors', 'asc', EntryType.WORKFLOW)).toEqual(-1);
+    expect(searchService.compareAttributes(a, b, 'all_authors', 'desc', EntryType.WORKFLOW)).toEqual(1);
+    expect(searchService.compareAttributes(b, c, 'all_authors', 'asc', EntryType.WORKFLOW)).toEqual(-1);
+    expect(searchService.compareAttributes(b, c, 'all_authors', 'desc', EntryType.WORKFLOW)).toEqual(1);
+    expect(searchService.compareAttributes(a, c, 'descriptorType', 'asc', EntryType.WORKFLOW)).toEqual(-1);
+    expect(searchService.compareAttributes(a, b, 'descriptorType', 'desc', EntryType.WORKFLOW)).toEqual(-0);
+    expect(searchService.compareAttributes(a, b, 'starredUsers', 'asc', EntryType.WORKFLOW)).toEqual(-1);
+    expect(searchService.compareAttributes(a, b, 'name', 'asc', EntryType.WORKFLOW)).toEqual(-1);
+    expect(searchService.compareAttributes(a, b, 'name', 'desc', EntryType.WORKFLOW)).toEqual(1);
+    expect(searchService.compareAttributes(b, c, 'name', 'asc', EntryType.WORKFLOW)).toEqual(-1);
+    expect(searchService.compareAttributes(b, c, 'name', 'desc', EntryType.WORKFLOW)).toEqual(-1);
   }));
 });
