+++ src/app/search/state/search.service.ts
@@ -322,32 +322,14 @@ export class SearchService {
     this.setSearchText(suggestTerm);
   }
 
-  setShowTagCloud(entryType: 'tool' | 'workflow' | 'notebook') {
-    if (entryType === 'tool') {
-      const showTagCloud: boolean = this.searchQuery.getValue().showToolTagCloud;
-      this.searchStore.update((state) => {
-        return {
-          ...state,
-          showToolTagCloud: !showTagCloud,
-        };
-      });
-    } else if (entryType === 'workflow') {
-      const showTagCloud: boolean = this.searchQuery.getValue().showWorkflowTagCloud;
-      this.searchStore.update((state) => {
-        return {
-          ...state,
-          showWorkflowTagCloud: !showTagCloud,
-        };
-      });
-    } else {
-      const showTagCloud: boolean = this.searchQuery.getValue().showNotebookTagCloud;
-      this.searchStore.update((state) => {
-        return {
-          ...state,
-          showNotebookTagCloud: !showTagCloud,
-        };
-      });
-    }
+  setShowTagCloud(entryType: EntryType) {
+    this.searchStore.update((state) => {
+      return {
+        ...state,
+        showTagCloud: !(this.searchQuery.getValue().showTagCloud && this.searchQuery.getValue().currentEntryType === entryType),
+        currentEntryType: entryType,
+      };
+    });
   }
 
   /**
